fix(account_settings): Avoid mixing up FSReadFile / FSCloseFile patches
This copy-and-paste error was a culprit of the bug described in issue #55.
It made the Account Settings applet on Wii U softlock during its initialisation, forcing the user to power off the system.

The applet now works again. However, trying to access it a second time without rebooting still freezes the Wii U, which is another regression listed in issue #55.